+++ src/graphql-config/apollo-config-factory.service.ts
@@ -1,18 +1,19 @@
 import { makeExecutableSchema } from '@graphql-tools/schema'
 import { SchemaGeneratorService } from './schema-generator.service'
 import { ApolloContext } from '../app/api.service'
-import { Config } from 'apollo-server-core/src/types'
+import { ApolloServerOptions } from '@apollo/server'
 
 export class ApolloConfigFactoryService {
   constructor(private readonly schemaGenerator: SchemaGeneratorService) {}
 
-  async createGqlOptions(context: ApolloContext): Promise<Config> {
+  async createGqlOptions(
+    context: ApolloContext,
+  ): Promise<ApolloServerOptions<ApolloContext>> {
     const schemaDefinition = await this.schemaGenerator.generateSchema(context)
     const schema = makeExecutableSchema(schemaDefinition)
 
     return {
       schema: schema,
-      playground: true,
-    } as Config
+    } as ApolloServerOptions<ApolloContext>
   }
 }